# Use unboxed closures everywhere