--- a/src/rust-crypto/cryptoutil.rs
+++ b/src/rust-crypto/cryptoutil.rs
@@ -249,7 +249,7 @@ pub fn add_bytes_to_bits_tuple
 pub trait FixedBuffer {
     /// Input a vector of bytes. If the buffer becomes full, process it with the provided
     /// function and then clear the buffer.
-    fn input(&mut self, input: &[u8], func: &mut FnMut(&[u8]));
+    fn input<F: FnMut(&[u8])>(&mut self, input: &[u8], func: F);
 
     /// Reset the buffer.
     fn reset(&mut self);
@@ -280,7 +280,7 @@ pub trait FixedBuffer {
 
 macro_rules! impl_fixed_buffer( ($name:ident, $size:expr) => (
     impl FixedBuffer for $name {
-        fn input(&mut self, input: &[u8], func: &mut FnMut(&[u8])) {
+        fn input<F: FnMut(&[u8])>(&mut self, input: &[u8], mut func: F) {
             let mut i = 0;
 
             // FIXME: #6304 - This local variable shouldn't be necessary.
@@ -403,11 +403,11 @@ pub trait StandardPadding {
     /// and is guaranteed to have exactly rem remaining bytes when it returns. If there are not at
     /// least rem bytes available, the buffer will be zero padded, processed, cleared, and then
     /// filled with zeros again until only rem bytes are remaining.
-    fn standard_padding(&mut self, rem: uint, func: &mut FnMut(&[u8]));
+    fn standard_padding<F: FnMut(&[u8])>(&mut self, rem: uint, func: F);
 }
 
 impl <T: FixedBuffer> StandardPadding for T {
-    fn standard_padding(&mut self, rem: uint, func: &mut FnMut(&[u8])) {
+    fn standard_padding<F: FnMut(&[u8])>(&mut self, rem: uint, mut func: F) {
         let size = self.size();
 
         self.next(1)[0] = 128;