Use unboxed closures everywhere that closures are used. This basically just effects symmetric cipher blockmode tests and the digests. The digests seem to see a very minor performance improvement:

before:
```
test sha1::bench::sha1_10                         ... bench:        48 ns/iter (+/- 1) = 208 MB/s
test sha1::bench::sha1_1k                         ... bench:      3986 ns/iter (+/- 375) = 256 MB/s
test sha1::bench::sha1_64k                        ... bench:    253095 ns/iter (+/- 2086) = 258 MB/s
```

after:
```
test sha1::bench::sha1_10                         ... bench:        46 ns/iter (+/- 0) = 217 MB/s
test sha1::bench::sha1_1k                         ... bench:      3924 ns/iter (+/- 41) = 260 MB/s
test sha1::bench::sha1_64k                        ... bench:    250463 ns/iter (+/- 3073) = 261 MB/s
```